refactor: class 정리
