+++ src/main/java/persistence/sql/ddl/EntityClassValidator.java
@@ -1,0 +1,32 @@
+package persistence.sql.ddl;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+import java.lang.reflect.Field;
+import java.util.Arrays;
+import java.util.List;
+
+public class EntityClassValidator {
+    private EntityClassValidator() {
+    }
+
+    static void validate(Class<?> clazz) {
+        if (!clazz.isAnnotationPresent(Entity.class)) {
+            throw new IllegalArgumentException("Entity must be annotated with @Entity");
+        }
+
+        validateHasId(clazz.getDeclaredFields());
+    }
+
+    static void validateHasId(Field[] fields) {
+        List<Field> idFields = Arrays.stream(fields)
+                .filter(field ->
+                        field.isAnnotationPresent(Id.class)
+                ).toList();
+
+        if (idFields.size() != 1) {
+            throw new IllegalArgumentException("Entity must have exactly one field annotated with @Id");
+        }
+    }
+}
